requirements.txt
numpy==1.23.1
opencv-python==4.6.0.66
python-dotenv==0.20.0
pandas>=1.4.3
Pillow>=8.4.0
tqdm>=4.64.0
keras>=2.9.0
tensorflow>=2.9.1
slackclient>=2.9.4
mtcnn>=0.8.10.1
opencv-python>=4.6.0.66
mediapipe>=0.8.10.1
gdown>=4.5.1
tkinter
